Добавляет компоненты AvatarBox.tsx, ChangeAvatar.tsx. Исправляет баг стандартного значения при клике на кнопку цели в случае не изменения ползунка value.
