+++ src/components/MainMenu/MainMenu.style.ts
@@ -2,14 +2,6 @@ import { createStyles, makeStyles, Theme } from '@material-ui/core/styles';
 
 export default makeStyles((theme: Theme) =>
   createStyles({
-    avatar: {
-      width: theme.spacing(8),
-      height: theme.spacing(8),
-    },
-    typography: {
-      background: 'none',
-      marginTop: theme.spacing(2),
-    },
     gridSideBar: {
       height: '100%',
     },
@@ -18,12 +10,6 @@ export default makeStyles((theme: Theme) =>
     },
     list: {
       flexGrow: 6,
-    },
-    gridAvatar: {
-      flexGrow: 1,
-    },
-    input: {
-      display: 'none',
     },
     link: {
       color: 'black',
