+++ src/router/Router.tsx
@@ -6,6 +6,7 @@ import LoginPage from '../pages/login';
 import RegistrationPage from '../pages/registration';
 import TasksPage from '../pages/tasks/TasksPage';
 import NewTaskPage from '../pages/tasks/add/NewTaskPage';
+import ChangeAvatar from '../components/ChangeAvatar/ChangeAvatar';
 
 // @ts-ignore
 const Router = ({ navbar: NavBar, template: Template }) => (
@@ -34,7 +35,7 @@ const Router = ({ navbar: NavBar, template: Template }) => (
         <TasksPage />
       </Route>
       <Route path="/goals">
-        <div>dsjsdgkj</div>
+        <ChangeAvatar />
       </Route>
     </Switch>
   </BrowserRouter>
